Update README.md
In 1965, a new simple game was invented, which is played on a
badminton court: pickleball. It is very similar to badminton, tennis,
and table tennis.Every body wants to know why is pickleball so
popular?
As a beginner, the player must know how to score accurately. It is
said that it is not so easy to learn how to score in pickleball. Have
a look at the popular and fastest growing sport and know what is
pickleball? But sometimes it is not true: many players forget the
score and get confused.
This article will help you remember the score, both in doubles and
singles. This article will guide players in learning to score in
pickleball